Add bitcoin mempool class
This commit adds a class which fetches unconfirmed transactions from a bitcoin mainnet node's mempool.
It is currently not optimized, which may cause it to use excessive amounts of memory or time.
Additionally, there are way too many transactions in a mempool than will fit in any computer's memory, when they are processed.
Work is underway to create a data storage for the mempool transactions, located on the disk.